ToDo-Project/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
beautifulsoup4==4.11.1
dj-database-url==0.5.0
Django==3.2.6
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap-form==3.4
django-bootstrap4==22.2
django-crispy-forms==1.14.0
django-filter==22.1
gunicorn==20.1.0
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.2
Unipath==1.1
whitenoise==5.3.0